# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Учебный продакшн-подобный сервис учёта клиентов.\n\nКлючевая механика: обычный пользователь не редактирует карточку\nклиента напрямую — он ПРЕДЛАГАЕТ правку, которая применяется\nтолько после одобрения администратором.\n\nДемо-учётки (заводятся миграцией 00004):\nadmin@example.com / admin12345 — роль admin\nuser@example.com / user12345 — роль user",
        "title": "Clients API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/clients": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Доступно только администратору.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Создать клиента",
                "parameters": [
                    {
                        "description": "Данные клиента",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.CreateClientRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.ClientResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Недостаточно прав",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Email уже занят другим клиентом",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/clients/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Доступно только администратору. PUT означает ПОЛНУЮ ЗАМЕНУ:\nнепереданные необязательные поля будут очищены.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Изменить клиента",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID клиента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные клиента",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.UpdateClientRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.ClientResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Мягкое удаление: строка остаётся в базе с меткой deleted_at,\nно исчезает из всех выдач.",
                "tags": [
                    "admin"
                ],
                "summary": "Удалить клиента",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID клиента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Удалено, тела ответа нет"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Уже удалён или не существует",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/clients": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Постраничная выдача с поиском по имени и компании,\nфильтром по статусу и сортировкой.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Список клиентов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Поиск по имени или компании",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "active",
                            "archived"
                        ],
                        "type": "string",
                        "description": "Фильтр по статусу",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Записей на страницу (1..100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Сколько пропустить",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "created_at",
                            "updated_at",
                            "full_name"
                        ],
                        "type": "string",
                        "default": "created_at",
                        "description": "Поле сортировки",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "Направление",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.ListResponse-github_com_vitikevich-landau_clients-api_internal_model_ClientResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/clients/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clients"
                ],
                "summary": "Карточка клиента",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID клиента",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.ClientResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный UUID",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Клиент не найден или удалён",
                        "schema": {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_vitikevich-landau_clients-api_internal_apperr.Code": {
            "type": "string",
            "enum": [
                "validation_error",
                "unauthorized",
                "forbidden",
                "not_found",
                "method_not_allowed",
                "conflict",
                "timeout",
                "internal_error"
            ],
            "x-enum-comments": {
                "CodeConflict": "нарушение уникальности / бизнес-правила",
                "CodeForbidden": "токен есть, но прав не хватает",
                "CodeInternal": "всё остальное, наша вина",
                "CodeMethodNotAllowed": "путь есть, метод не тот",
                "CodeNotFound": "объект не существует",
                "CodeTimeout": "не уложились в отведённое время",
                "CodeUnauthorized": "нет токена или он невалиден",
                "CodeValidation": "невалидные входные данные"
            },
            "x-enum-descriptions": [
                "невалидные входные данные",
                "нет токена или он невалиден",
                "токен есть, но прав не хватает",
                "объект не существует",
                "путь есть, метод не тот",
                "нарушение уникальности / бизнес-правила",
                "не уложились в отведённое время",
                "всё остальное, наша вина"
            ],
            "x-enum-varnames": [
                "CodeValidation",
                "CodeUnauthorized",
                "CodeForbidden",
                "CodeNotFound",
                "CodeMethodNotAllowed",
                "CodeConflict",
                "CodeTimeout",
                "CodeInternal"
            ]
        },
...
        "github_com_vitikevich-landau_clients-api_internal_model.ClientResponse": {
            "type": "object",
            "properties": {
                "company": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.ClientStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "github_com_vitikevich-landau_clients-api_internal_model.ClientStatus": {
            "type": "string",
            "enum": [
                "active",
                "archived"
            ],
            "x-enum-comments": {
                "ClientStatusActive": "работаем с клиентом",
                "ClientStatusArchived": "в архиве, но не удалён"
            },
            "x-enum-descriptions": [
                "работаем с клиентом",
                "в архиве, но не удалён"
            ],
            "x-enum-varnames": [
                "ClientStatusActive",
                "ClientStatusArchived"
            ]
        },
        "github_com_vitikevich-landau_clients-api_internal_model.CreateClientRequest": {
            "type": "object",
            "required": [
                "full_name"
            ],
            "properties": {
                "company": {
                    "type": "string",
                    "maxLength": 200
                },
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "full_name": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1
                },
                "notes": {
                    "type": "string",
                    "maxLength": 2000
                },
                "phone": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                },
                "status": {
                    "description": "Если статус не передали, применится ClientStatusActive.",
                    "enum": [
                        "active",
                        "archived"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.ClientStatus"
                        }
                    ]
                }
            }
        },
        "github_com_vitikevich-landau_clients-api_internal_model.ListResponse-github_com_vitikevich-landau_clients-api_internal_model_ClientResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.ClientResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.Pagination"
                }
            }
        },
...
        "github_com_vitikevich-landau_clients-api_internal_model.Pagination": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_vitikevich-landau_clients-api_internal_model.UpdateClientRequest": {
            "type": "object",
            "required": [
                "full_name"
            ],
            "properties": {
                "company": {
                    "type": "string",
                    "maxLength": 200
                },
                "email": {
                    "type": "string",
                    "maxLength": 255
                },
                "full_name": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 1
                },
                "notes": {
                    "type": "string",
                    "maxLength": 2000
                },
                "phone": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                },
                "status": {
                    "enum": [
                        "active",
                        "archived"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_model.ClientStatus"
                        }
                    ]
                }
            }
        },
...
        "github_com_vitikevich-landau_clients-api_internal_response.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code — машиночитаемый код: not_found, validation_error, ...\nКлиент принимает решения ПО КОДУ, а не по тексту:\nтекст можно переписать или перевести, код — контракт.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_apperr.Code"
                        }
                    ]
                },
                "details": {
                    "description": "Details — пояснения по конкретным полям при ошибке валидации.\nomitempty: в остальных случаях ключа в JSON просто не будет.",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "description": "Message — человекочитаемое описание. Безопасное: без имён таблиц,\nSQL и стектрейсов.",
                    "type": "string"
                },
                "request_id": {
                    "description": "RequestID — чтобы пользователь мог назвать его в поддержке,\nа мы по нему нашли весь запрос в логах.",
                    "type": "string"
                }
            }
        },
        "github_com_vitikevich-landau_clients-api_internal_response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_vitikevich-landau_clients-api_internal_response.ErrorBody"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите: Bearer \u003cваш JWT\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}